--- a/webpack.config.js
+++ b/webpack.config.js
@@ -9,6 +9,22 @@ const MomentLocalesPlugin = require('moment-locales-webpack-plugin');
 const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 const safePostCssParser = require('postcss-safe-parser');
 
+
+const prodPlugins = (isDev) => {
+  if (isDev) {
+    return [];
+  }
+
+  return [
+    new WorkboxPlugin.GenerateSW({
+      clientsClaim: true,
+      skipWaiting: true,
+      maximumFileSizeToCacheInBytes: 7340032,
+      exclude: ['build-meta.json', /\.map$/]
+    })
+  ]
+}
+
 module.exports = (env, argv) => {
   const mode = argv.mode || "development";
   const isDev = mode !== "production";
@@ -23,13 +39,14 @@ module.exports = (env, argv) => {
     },
     output: {
       path: path.resolve(__dirname, "dist"),
+      pathinfo: false,
       filename: isDev ? "js/bundle.[hash].js" : "js/bundle.prod.[hash].js",
       chunkFilename: "[name].[chunkhash].chunk.js",
       publicPath: "/",
     },
     optimization: {
       moduleIds: "hashed",
-      splitChunks: {
+      splitChunks: isDev? false : {
         cacheGroups: {
           commons: {
             test: /[\\/]node_modules[\\/]/,
@@ -65,7 +82,11 @@ module.exports = (env, argv) => {
       rules: [
         {
           test: /\.(ts|tsx)$/,
+          include: path.resolve(__dirname, 'src'),
           loader: "ts-loader",
+          options: {
+            transpileOnly: true
+          }
         },
         {
           enforce: "pre",
@@ -122,7 +143,7 @@ module.exports = (env, argv) => {
       new HtmlWebpackPlugin({
         template: path.resolve(__dirname, "src", "index.html"),
         title: "Coronasafe Care",
-        minify: {
+        minify: isDev ? false : {
           removeComments: true,
           collapseWhitespace: true,
           removeRedundantAttributes: true,
@@ -152,12 +173,7 @@ module.exports = (env, argv) => {
           ? "css/[name][hash].bundle.css"
           : "css/[name][hash].prod.bundle.css",
       }),
-      new WorkboxPlugin.GenerateSW({
-        clientsClaim: true,
-        skipWaiting: true,
-        maximumFileSizeToCacheInBytes: 7340032,
-        exclude: ['build-meta.json', /\.map$/]
-      })
+      ...prodPlugins(isDev)
     ],
     performance: false,
   };