A tiny effort to reduce build times.

Ran from my lap:
 old initial dev build time: 65320ms
 old rebuild time: 31148ms

 new initial dev build time: 65110ms
 new rebuild time: 26001ms

While initial build time remains the same(due to changing source map options to "eval-cheap-module-source-map"),
rebuild time is faster by about 5seconds.

Some other areas to explore are:
-   https://github.com/TypeStrong/ts-loader#faster-builds
-   using just reason.js or typescript but not both
-   using SplitChunksPlugin plugin (https://webpack.js.org/guides/build-performance/#smaller--faster)